Adding HostKeyCallback to sshConfig. Fix command.
Added HostKeyCallback to sshConfig. Right now using the
ssh.InsecureIgnoreHost() callback. This returns a function that
in turn returns nil. Note: returning nil would result in acceptance
of any host key. Hence, we need to change this to a callback of our
own implementation.

Fixed the command to run the throttling script. There were spacing
issues. Also, using strings.Join(...) instead of adding spaces in
the string (which is error prone).